requirements.txt
aiogram
aiofiles
aiohttp
aiosignal
aniso8601
annotated-types
attrs
blinker
certifi
charset-normalizer
click
colorama
Flask
Flask-Mail
Flask-RESTful
frozenlist
googletrans
greenlet
h11
h2
hpack
hstspreload
httpcore
httpx
hyperframe
idna
itsdangerous
Jinja2
magic-filter
MarkupSafe
multidict
pillow
pydantic
pydantic_core
pytz
requests
rfc3986
setuptools
six
slack_sdk
sniffio
SQLAlchemy
typing_extensions
urllib3
Werkzeug
wheel
yarl
